Fixes for Swift 5; added new music

+++ TukTuk/Animations/FaceBalls.swift
@@ -32,7 +32,7 @@ class FaceBalls: Animation {
             (i, imageName) in
 
             DispatchQueue.main.asyncAfter(deadline: .now() + 0.25 * Double(i)) {
-                let radius: CGFloat = 30 + [10, 20, 30].randomElement()!
+                let radius: CGFloat = 30 + CGFloat([10, 20, 30].randomElement()!)
                 let face = SKShapeNode(circleOfRadius: radius)
                 face.lineWidth = 1
                 face.fillColor = .white
@@ -42,9 +42,9 @@ class FaceBalls: Animation {
                 let sprite: SKSpriteNode = SKSpriteNode(texture: texture)
                 sprite.physicsBody = SKPhysicsBody(circleOfRadius: radius)
                 sprite.physicsBody?.usesPreciseCollisionDetection = true
-                sprite.physicsBody?.restitution = 0.9 + [-0.05, 0, 0.05].randomElement()!
-                sprite.position = CGPoint(x: scene.frame.width * 0.4 + [-40, -20, 0, 20, 40].randomElement()!, y: scene.frame.height)
-                sprite.zRotation = 45.0 * [-1, 1].randomElement()!
+                sprite.physicsBody?.restitution = 0.9 + CGFloat([-0.05, 0, 0.05].randomElement()!)
+                sprite.position = CGPoint(x: scene.frame.width * 0.4 + CGFloat([-40, -20, 0, 20, 40].randomElement()!), y: scene.frame.height)
+                sprite.zRotation = CGFloat(45.0 * [-1, 1].randomElement()!)
                 scene.addChild(sprite)
             }
         }
